+++ src/components/admin/games/index.ts
@@ -2,3 +2,4 @@ export { GameContentModal } from "./GameContentModal";
 export { HangmanContentEditor } from "./HangmanContentEditor";
 export { WordSearchContentEditor } from "./WordSearchContentEditor";
 export { MemoryContentEditor } from "./MemoryContentEditor";
+export { QuizContentEditor } from "./QuizContentEditor";